added restricted pages and login/logout

+++ chordity/base/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -16,6 +19,9 @@ from .forms import RoomForm
 
 def loginPage(request):
 
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -23,9 +29,22 @@ def loginPage(request):
         try:
             user = User.objects.get(username=username)
         except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None: 
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password does not exist.')
         
     context = {}
     return render(request, "base/login_register.html", context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 
 def home(request):
@@ -55,7 +74,7 @@ def room(request, pk):
 
     return render(request, "base/room.html", context)
 
-
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
 
@@ -69,9 +88,13 @@ def createRoom(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('User does not have permissions to access this.')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -83,8 +106,13 @@ def updateRoom(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('User does not have permissions to access this.')
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
